Move provider issue generation to breeze (#28352)
The exercise where another PMC performed a release process had
proven useful to show the value of Breeze in release management
commands - there were no problems with running the Breeze commands,
but for scripts put elsewhere (docs building, dev) it was a bit
problematic because each required own virtualenv.

This is the first PR to simplify that - one less venv for the
issue generation process - now it is moved to Breeze so no more
setup is needed.

(cherry picked from commit fd5846d256b6d269b160deb8df67cd3d914188e0)
