+++ dev/breeze/setup.cfg
@@ -58,12 +58,14 @@ install_requires =
     filelock
     inputimeout
     importlib-metadata>=4.4; python_version < "3.8"
+    jinja2
     pendulum
     pre-commit
     psutil
     pytest
     pytest-xdist
     pyyaml
+    PyGithub
     requests
     rich>=12.6.0
     rich-click>=1.5
