+++ images/breeze/output-commands-hash.txt
@@ -36,21 +36,22 @@ prod-image:pull:e3c89dd908fc44adf6e159c2950ebdd0
 prod-image:verify:31bc5efada1d70a0a31990025db1a093
 prod-image:a4013428dc7f71a1defc3778d2efe3dc
 release-management:generate-constraints:ae30d6ad49a1b2c15b61cb29080fd957
+release-management:generate-issue-content:24218438f9e85e7c92258aadebbb19de
 release-management:prepare-airflow-package:3ac14ea6d2b09614959c0ec4fd564789
 release-management:prepare-provider-documentation:3fe5ead9887c518d1b397d1103dc0025
 release-management:prepare-provider-packages:40144cb01afc56f6a4f92d9e117e546e
 release-management:release-prod-images:c9bc40938e0efad49e51ef66e83f9527
-release-management:verify-provider-packages:8d3c6362657d76bb3cd064fed5596e84
-release-management:c4e840ca22b11a431f5bc7b118dc061f
+release-management:verify-provider-packages:88bd609aff6d09d52ab8d80d6e055e7b
+release-management:eafd13da512a5a7c0fb1499cf7ff1d63
 setup:autocomplete:03343478bf1d0cf9c101d454cdb63b68
 setup:config:3ffcd35dd24b486ddf1d08b797e3d017
-setup:regenerate-command-images:255746830d7b5d1337d13b8e101f7f83
+setup:regenerate-command-images:ab2d83c339fa3a42b0c819b6b6cc88ae
 setup:self-upgrade:d02f70c7a230eae3463ceec2056b63fa
 setup:version:123b462a421884dc2320ffc5e54b2478
-setup:2e9e4ab1729c5420b7a2b78cbee7539a
-shell:affbf6f7f469408d0af47f75c6a38f6c
-start-airflow:109728919a0dd5c5ff5640ae86ba9e90
-static-checks:7a39e28c87fbca0a9fae0ebfe1591b71
+setup:fbabee281b69f818091d780b24bd815a
+shell:76e0f530b7af514a2aad3032b6516c46
+start-airflow:06d4aeb5f1b65f6b975f3f915558d0b3
+static-checks:f45ad432bdc47a2256fdb0277b19d816
 stop:8969537ccdd799f692ccb8600a7bbed6
 testing:docker-compose-tests:b86c044b24138af0659a05ed6331576c
 testing:helm-tests:94a442e7f3f63b34c4831a84d165690a
